Remove unneeded imports, move functionality from App into Main component.
