+++ my-app/components/SplashScreen.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import { View, Text, StyleSheet, Image } from 'react-native';
 
 const SplashScreen = ({ navigation }) => {
